triage_issues/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
fasttext==0.9.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
lxml==5.1.0
numpy==1.26.4
openai==1.13.3
pybind11==2.11.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
rapidfuzz==3.6.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
thefuzz==0.22.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
